[GRPC] Support abstract socket and an existing socket's fd.
This CL adds special handling for Android Studio Profilers.

(1) Supports Unix abstract domain socket if the input address starts
with prefix "unix:@". For example, "unix:@hidden" indicates an
abstract socket of name 'hidden'. Note the name of the abstract
socket doesn't include the '@' immediately after "unix:".
Abstract socket supports both server and client modes.

(2) Supports a GRPC client to use an existing socket if the file
descriptor as an integer is provided in the format of "unix:&FD".
For example, "unix:&123" indicates a socket of file descriptor
'123' exists and is going to be used. Existing socket's file
descriptor is supported only in clients.

Bug: 35672396

Test: manual
<enable.jvmti.profiling=true/false> X <O and older devices>

Change-Id: I0839558371e843cf5ad0d2391f10c501c9aa1541