Support for the emulator build system.
This adds the cmake templates that work together with the cross build
system that the emulator uses.

By default the tests will not be build and run:

- The depend on gflags & newer benchmark that we do not
  have in our repo.
- They depend on their own test runner.
- They are bit a flaky
- They cause firewall issues on mac.

Note: We shoud likely move these to a seperate file to reduce cmake
configuration times.

If you want to regenerate the cmake files you have to exectue the meta
build generator as follows:

$ pip install --user prettytable Mako pyaml dateutils --upgrade
$ tools/buildgen/generate_projects.sh --templates \
        templates/emulator/CMakeLists.txt.template \
        templates/emulator/plugins/CMakeLists.txt.template \

Change-Id: I3124db6f8191e200789c6f5f32600f4136d30327