[GRPC] Support abstract socket and an connected fd
This CL is an re-implementation of commit 35dd92ca0c464fea9cf41ba7e49e9ab34f361366
while gRPC is being upgraded to 1.22.0.

This CL adds special handling for Android Studio Profilers and
Transport Pipeline in general.

(1) Supports Unix abstract domain socket if the input address starts
with prefix "unix:@". For example, "unix:@hidden" indicates an
abstract socket of name 'hidden'. Note the name of the abstract
socket doesn't include the '@' immediately after "unix:".
Abstract socket supports both server and client modes.

(2) Supports a GRPC client to use an existing socket if the file
descriptor as an integer is provided in the format of "unix:&FD".
For example, "unix:&123" indicates a socket of file descriptor
'123' exists and is going to be used. Existing socket's file
descriptor is supported only in clients.

Bug: 138846475
Test: will add unit tests in a follow up CL
Change-Id: I433df6eae990253b713b10f42d5acfe920e102f1