Disable packages we do not need
gRPC contains a large set of features that we do not need, but require
a large set of dependencies if we would support them.

We do not want to bring in envoy, google cloud support or ios support.
The unused functionality has been commented out, so we can hopefully
merge down changes relatively easily.

Change-Id: I38c4e4e62e0f291a49cfc4e70ae35cd5947478d6